+++ src/main/gui/SamplePoint.cpp
@@ -1,6 +1,6 @@
 #include "SamplePoint.h"
 
-SamplePoint::SamplePoint(CupuacuState *state, const uint8_t channelIndexToUse, const size_t sampleIndexToUse) :
+SamplePoint::SamplePoint(CupuacuState *state, const uint8_t channelIndexToUse, const int64_t sampleIndexToUse) :
     Component(state, "Sample point idx " + std::to_string(sampleIndexToUse)), sampleIndex(sampleIndexToUse), channelIndex(channelIndexToUse)
 {
 }
